Refactored structs into a single structs package
