+++ golang/entry/entry.go
@@ -1,0 +1,13 @@
+// Package entry provides methods for endpoints for registration and login
+package entry
+
+import (
+	"context"
+	"log"
+	"os"
+	"time"
+
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
